requirements.txt
absl-py
acres
antspyx
astunparse
cachetools
certifi
charset-normalizer
ci-info
click
contourpy
cycler
etelemetry
filelock
flatbuffers
fonttools
fsspec
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
imageio
importlib_metadata
importlib_resources
isodate
Jinja2
joblib
keras
kiwisolver
lazy_loader
libclang
looseversion
lxml
Markdown
MarkupSafe
matplotlib
mpmath
networkx
neurite
nibabel
nipype
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
opt_einsum
packaging
pandas
patsy
pillow
protobuf
prov
puremagic
pyasn1
pyasn1_modules
pydot
pyparsing
pystrum
python-dateutil
pytz
PyYAML
rdflib
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
simplejson
six
statsmodels
surfa
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tifffile
torch
torchaudio
torchvision
tqdm
traits
triton
tzdata
urllib3
voxelmorph
webcolors
Werkzeug
wrapt
xxhash
zipp
